requirement.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_tflow_select=2.1.0=gpu
absl-py=0.8.0=py36_0
astor=0.8.0=py36_0
attrs=19.3.0=py_0
backcall=0.1.0=py_0
blas=1.0=mkl
bleach=3.1.0=py_0
c-ares=1.15.0=h7b6447c_1001
ca-certificates=2019.10.16=0
cairo=1.14.12=h8948797_3
certifi=2019.9.11=py36_0
chardet=3.0.4=pypi_0
cloudpickle=1.2.2=pypi_0
cudatoolkit=10.1.168=0
cudnn=7.6.0=cuda10.1_0
cupti=10.1.168=0
cycler=0.10.0=pypi_0
decorator=4.4.0=py_0
defusedxml=0.6.0=py_0
dill=0.3.0=py36_0
entrypoints=0.3=py36_1000
expat=2.2.6=he6710b0_0
fontconfig=2.13.0=h9420a91_0
freetype=2.9.1=h8a8886c_1
fribidi=1.0.5=h7b6447c_0
gast=0.3.2=py_0
glib=2.56.2=hd408876_0
google-pasta=0.1.7=py_0
graphite2=1.3.13=h23475e2_0
graphviz=2.40.1=h21bd128_2
grpcio=1.16.1=py36hf8bcb03_1
h5py=2.9.0=py36h7918eee_0
harfbuzz=1.8.8=hffaf4a1_0
hdf5=1.10.4=hb1b8bf9_0
icu=58.2=h9c2bf20_1
idna=2.8=pypi_0
imageio=2.5.0=pypi_0
importlib_metadata=0.23=py36_0
intel-openmp=2019.4=243
ipykernel=5.1.3=py36h5ca1d4c_0
ipython=7.8.0=py36h5ca1d4c_0
ipython_genutils=0.2.0=py_1
jedi=0.15.1=py36_0
jinja2=2.10.3=py_0
joblib=0.14.0=pypi_0
jpeg=9b=h024ee3a_2
json5=0.8.5=py_0
jsonschema=3.1.1=py36_0
jupyter_client=5.3.3=py36_1
jupyter_core=4.5.0=py_0
jupyterlab=1.1.4=py_0
jupyterlab_server=1.0.6=py_0
keras-applications=1.0.8=py_0
keras-base=2.2.4=py36_0
keras-gpu=2.2.4=0
keras-preprocessing=1.1.0=py_1
kiwisolver=1.1.0=pypi_0
libedit=3.1.20181209=hc058e9b_0
libffi=3.2.1=hd88cf55_4
libgcc-ng=9.1.0=hdf63c60_0
libgfortran-ng=7.3.0=hdf63c60_0
libpng=1.6.37=hbc83047_0
libprotobuf=3.9.2=hd408876_0
libsodium=1.0.17=h516909a_0
libstdcxx-ng=9.1.0=hdf63c60_0
libtiff=4.0.10=h2733197_2
libuuid=1.0.3=h1bed415_2
libxcb=1.13=h1bed415_1
libxml2=2.9.9=hea5a465_1
markdown=3.1.1=py36_0
markupsafe=1.1.1=py36h14c3975_0
matplotlib=3.1.1=pypi_0
mistune=0.8.4=py36h14c3975_1000
mkl=2019.4=243
mkl-service=2.3.0=py36he904b0f_0
mkl_fft=1.0.14=py36ha843d7b_0
mkl_random=1.1.0=py36hd6b4f25_0
more-itertools=7.2.0=py_0
nbconvert=5.6.1=py36_0
nbformat=4.4.0=py_1
ncurses=6.1=he6710b0_1
networkx=2.4=pypi_0
notebook=6.0.1=py36_0
numpy=1.16.5=pypi_0
numpy-base=1.17.2=py36hde5b4d6_0
openssl=1.1.1=h7b6447c_0
pandoc=2.7.3=0
pandocfilters=1.4.2=py_1
pango=1.42.4=h049681c_0
parso=0.5.1=py_0
pcre=8.43=he6710b0_0
pexpect=4.7.0=py36_0
pickleshare=0.7.5=py36_1000
pillow=6.2.1=pypi_0
pip=19.2.3=py36_0
pixman=0.38.0=h7b6447c_0
progressbar2=3.39.3=pypi_0
prometheus_client=0.7.1=py_0
prompt_toolkit=2.0.10=py_0
protobuf=3.9.2=py36he6710b0_0
ptyprocess=0.6.0=py_1001
pydot=1.4.1=pypi_0
pygments=2.4.2=py_0
pyparsing=2.4.2=pypi_0
pyrsistent=0.15.4=py36h516909a_0
python=3.6.9=h265db76_0
python-dateutil=2.8.0=py_0
python-graphviz=0.13=pypi_0
python-utils=2.3.0=pypi_0
pywavelets=1.1.1=pypi_0
pyyaml=5.1.2=py36h7b6447c_0
pyzmq=18.1.0=py36h1768529_0
readline=7.0=h7b6447c_5
requests=2.21.0=pypi_0
scikit-image=0.15.0=pypi_0
scikit-learn=0.21.0=pypi_0
scipy=1.2.1=pypi_0
send2trash=1.5.0=py_0
setuptools=41.4.0=py36_0
six=1.12.0=py36_0
sklearn=0.0=pypi_0
sqlite=3.30.0=h7b6447c_0
tensorboard=1.14.0=py36hf484d3e_0
tensorflow=1.14.0=gpu_py36h3fb9ad6_0
tensorflow-base=1.14.0=gpu_py36he45bfe2_0
tensorflow-estimator=1.14.0=py_0
tensorflow-gpu=1.14.0=h0d30ee6_0
tensorlayer=2.1.0=pypi_0
termcolor=1.1.0=py36_1
terminado=0.8.2=py36_0
testpath=0.4.2=py_1001
tflearn=0.3.2=pypi_0
tk=8.6.8=hbc83047_0
tornado=6.0.3=py36h516909a_0
traitlets=4.3.3=py36_0
urllib3=1.24.3=pypi_0
wcwidth=0.1.7=py_1
webencodings=0.5.1=py_1
werkzeug=0.16.0=py_0
wheel=0.33.6=py36_0
wrapt=1.11.1=pypi_0
xz=5.2.4=h14c3975_4
yaml=0.1.7=had09818_2
zeromq=4.3.2=he1b5a44_2
zipp=0.6.0=py_0
zlib=1.2.11=h7b6447c_3
zstd=1.3.7=h0b5b093_0
